Writing to subdirectories named after discs.
Sometimes makemkvcon writes filenames such as title00.mkv. This poses a
problem if someone is ripping a TV series, since every disk will have
that sort of filenames. Everything in a flat directory will lead to
error if the user doesn't move files out of the output dir on every disc
ripped and this breaks systems with multiple optical drives.

Fixing this by adding DIRECTORY env variable. By default it's the disc's
name or UUID (if name is empty). Can be overridden by the user.