Splitting rip.sh into two files.
Script was getting large, but I didn't want to convert it to Python. So
I'm splitting it into two: env.sh and rip.sh. The former will handle
read-only on-source statements with a prepare() function that prepares
the container's environment.

Exposing DEVNAME, but not yet testing it.

Running groupmod instead of adding a new group inside the container.
Nothing else uses the existing cdrom group IDs.

Explicitly specifying device name for all commands that interface with
it (makemkvcon, blkid, eject, etc).