Exposing UMASK environment variable.
Allow users to specify a umask so files and directories are created with
those permissions.

My use case was I wanted MakeMKV to create group-writable files.
makemkvcon apparently hard-coded mode 644 so even when umask is applied
I couldn't get 664 on file creation. The solution is to "hack"
makemkvcon by intercepting the open(3) syscall to maximize permissions
and allow the umask to downgrade them instead. This is how "touch" does
it.