s/force_umask/wrappers/, gcc in CI, is_mkv().
Renamed force_umask to wrappers since it will do more in the future.

Running gcc in CI to test for compile errors there. Also compiles for me
and uploads it as an artifact.

Re-structured C code. Moved logic into is_mkv(). Using an init()
constructor instead of that if statement on every open() call.