Adding hook-post-title.sh support.
Patching the close() glibc function of makemkvcon to signal when an MKV
file is done ripping. Using a FIFO file to transmit data from makemkvcon
(run as the mkv user) to run.sh (run as root).

This lets me run ffmpeg right after an MKV file is done being written to
to strip out subtitles and other languages in my hook script (in another
branch) instead of waiting until the end of makemkvcon to process all
files.