Fix handling of background jobs in hook-post-title
When users start long-running processes in hook-post-title, rip.sh
should wait for them at the end of the run. However since the hook is
sourced in a background task that reads the FIFO, those children aren't
related to the main rip.sh script. The "wait" command doesn't wait for
grandchildren.

Fixing this by having makemkvcon run in the background while the FIFO is
read in the foreground. All background jobs from hook-post-title are
direct children of rip.sh, not grandchildren. Hence those jobs are now
waited on at the end of the run.